Start handler for sql and socket events
